move log handlers into separated file
